fix matching on K for hax

+++ src/kem/kyber/hash_functions.rs
@@ -26,7 +26,7 @@ pub(crate) enum XofState<const K: usize> {
 
 #[inline(always)]
 pub(crate) fn XOF_absorb<const K: usize>(input: [[u8; 34]; K]) -> XofState<K> {
-    match K {
+    match K as u8 {
         2 => {let mut state = digest::shake128_init_x2();
               digest::shake128_absorb_final_x2(&mut state, &input[0], &input[1]);
               XofState::X2(state)},
@@ -45,7 +45,7 @@ pub(crate) fn XOF_squeeze_three_blocks<const K: usize>(
     xof_state: XofState<K>,
 ) -> ([[u8; 168 * 3]; K],XofState<K>) {
     let mut output = [[0; 168 * 3]; K];
-    match (K, xof_state) {
+    match (K as u8, xof_state) {
         (2, XofState::X2(mut st)) => {
             let tmp = digest::shake128_squeeze_nblocks_x2::<504>(&mut st);
             output[0] = tmp[0];
@@ -76,7 +76,7 @@ pub(crate) fn XOF_squeeze_block<const K: usize>(
     xof_state: XofState<K>,
 ) -> ([[u8; 168]; K],XofState<K>) {
     let mut output = [[0; 168]; K];
-    match (K, xof_state) {
+    match (K as u8, xof_state) {
         (2, XofState::X2(mut st)) => {
             let tmp = digest::shake128_squeeze_nblocks_x2::<168>(&mut st);
             output[0] = tmp[0];
